refactor: remove github issues internal callback route
