+++ turbo/apps/api/src/signals/routes/__tests__/helpers/api-bdd-github.ts
@@ -52,8 +52,6 @@ export const GH_OAUTH_CLIENT_SECRET = "github-oauth-client-secret";
 
 const GITHUB_APP_ID = "123456";
 const DEFAULT_TEST_ORIGIN = "http://localhost:3000";
-const GITHUB_ISSUES_CALLBACK_PATH = "/api/internal/callbacks/github/issues";
-const GITHUB_ISSUES_CALLBACK_URL = `${DEFAULT_TEST_ORIGIN}${GITHUB_ISSUES_CALLBACK_PATH}`;
 
 type ComposeContent = z.infer<typeof agentComposeApiContentSchema>;
 
@@ -487,23 +485,6 @@ export function captureGithubIssueApi(
   };
 }
 
-/**
- * Forward internal GitHub-issues callback dispatches back into the app so
- * detached terminal transitions deliver their callbacks for real.
- */
-export function proxyGithubIssuesCallbackToApp(context: TestContext): void {
-  server.use(
-    http.post(GITHUB_ISSUES_CALLBACK_URL, async ({ request }) => {
-      const app = createApp({ signal: context.signal });
-      return await app.request(GITHUB_ISSUES_CALLBACK_PATH, {
-        method: "POST",
-        headers: request.headers,
-        body: await request.text(),
-      });
-    }),
-  );
-}
-
 export function buildLegacySignedState(args: {
   readonly userId: string;
   readonly composeId: string;
@@ -914,26 +895,6 @@ export function createGithubBddApi(context: TestContext) {
         }),
         [200],
       );
-    },
-
-    async requestGithubIssuesCallback(
-      rawBody: string,
-      headers: Record<string, string>,
-      statuses: readonly number[],
-    ): Promise<{ readonly status: number; readonly body: unknown }> {
-      const app = createApp({ signal: context.signal });
-      const response = await app.request(GITHUB_ISSUES_CALLBACK_PATH, {
-        method: "POST",
-        headers: { "content-type": "application/json", ...headers },
-        body: rawBody,
-      });
-      const body: unknown = await response.json();
-      if (!statuses.includes(response.status)) {
-        throw new Error(
-          `Expected GitHub issues callback status in [${statuses.join(", ")}], received ${response.status}: ${JSON.stringify(body)}`,
-        );
-      }
-      return { status: response.status, body };
     },
 
     /**
